+++ src/adapters/CanKit.Adapter.SocketCAN/Registers/SocketCanCoreRegister.cs
@@ -1,4 +1,3 @@
-using CanKit.Abstractions.Attributes;
 using CanKit.Abstractions.SPI.Factories;
 using CanKit.Abstractions.SPI.Providers;
 using CanKit.Abstractions.SPI.Registry.Core;
@@ -7,7 +6,6 @@ using CanKit.Core.Registry;
 
 namespace CanKit.Adapter.SocketCAN.Registers;
 
-[CanRegistryEntry(CanRegistryEntryKind.Adapter, "SocketCAN")]
 internal sealed class SocketCanCoreRegister : ICanRegisterFactory, ICanRegisterProviders, IRawRegisterEndpoint
 {